+++ packages/blocto-sdk/src/providers/ethereum.ts
@@ -16,10 +16,9 @@ import addSelfRemovableHandler from '../lib/addSelfRemovableHandler';
 import {
   setAccountStorage,
   getAccountStorage,
-  removeChainAddress,
-  getChainAddress,
-  setChainAddress,
-  KEY_SESSION,
+  getEvmAddress,
+  setEvmAddress,
+  removeAllEvmAddress,
 } from '../lib/storage';
 import responseSessionGuard, {
   ICustomError,
@@ -30,6 +29,8 @@ import {
   DEFAULT_APP_ID,
   ETH_SESSION_KEY_MAPPING,
   SDK_VERSION,
+  KEY_SESSION,
+  CHAIN,
 } from '../constants';
 import { isEmail } from '../lib/isEmail';
 import {
@@ -152,7 +153,11 @@ export default class EthereumProvider
       display_name,
       network_type,
       blocto_service_environment,
-    } = supportNetworkList[chainId];
+    } = supportNetworkList[chainId] ?? {};
+    if (!chain_id)
+      throw ethErrors.provider.unsupportedMethod(
+        `Get support chain failed: ${chainId} might not be supported yet.`
+      );
     const wallet_web_url =
       ETH_ENV_WALLET_SERVER_MAPPING[blocto_service_environment];
     this._blocto.switchableNetwork[chain_id] = {
@@ -172,6 +177,15 @@ export default class EthereumProvider
       parseChainId(existedSDK.chainId) !== parseChainId(this.chainId)
     ) {
       throw ethErrors.provider.chainDisconnected();
+    }
+  }
+
+  #onAccountChanged(event: MessageEvent): void {
+    if (
+      event.data?.type === 'BLOCTO_SDK:ACCOUNT_CHANGED' &&
+      event.data?.originChain !== CHAIN.ETHEREUM
+    ) {
+      this.handleDisconnect();
     }
   }
 
@@ -347,10 +361,13 @@ export default class EthereumProvider
       case 'wallet_switchEthereumChain': {
         return this.handleSwitchChain(payload?.params?.[0]?.chainId);
       }
+      case 'wallet_disconnect': {
+        return this.handleDisconnect();
+      }
     }
 
     // Method that requires user to be connected
-    if (!getChainAddress(sessionKey, blockchainName)) {
+    if (!getEvmAddress(sessionKey, blockchainName)) {
       const email = payload?.params?.[0];
       if (payload.method === 'eth_requestAccounts' && isEmail(email)) {
         await this.enable(email);
@@ -366,10 +383,10 @@ export default class EthereumProvider
           await this.fetchAccounts();
         // eslint-disable-next-line
         case 'eth_accounts':
-          result = getChainAddress(sessionKey, blockchainName);
+          result = getEvmAddress(sessionKey, blockchainName);
           break;
         case 'eth_coinbase': {
-          result = getChainAddress(sessionKey, blockchainName)?.[0];
+          result = getEvmAddress(sessionKey, blockchainName)?.[0];
           break;
         }
         case 'eth_signTypedData_v3':
@@ -378,11 +395,6 @@ export default class EthereumProvider
         case 'personal_sign':
         case 'eth_sign': {
           result = await this.handleSign(payload);
-          break;
-        }
-        case 'wallet_disconnect': {
-          this.handleDisconnect();
-          result = null;
           break;
         }
         case 'eth_sendTransaction':
@@ -542,7 +554,7 @@ export default class EthereumProvider
           method: 'wallet_switchEthereumChain',
           params: [{ chainId: this.chainId }],
         });
-        setChainAddress(sessionKey, blockchainName, [existedSDK.address]);
+        setEvmAddress(sessionKey, blockchainName, [existedSDK.address]);
       }
       return new Promise((resolve, reject) =>
         // add a small delay to make sure the network has been switched
@@ -550,7 +562,7 @@ export default class EthereumProvider
       );
     }
 
-    const address = getChainAddress(sessionKey, blockchainName);
+    const address = getEvmAddress(sessionKey, blockchainName);
     if (address) {
       return new Promise((resolve) => {
         resolve(address);
@@ -581,13 +593,19 @@ export default class EthereumProvider
                 sessionKey,
                 {
                   code: e.data.code,
-                  connected: true,
-                  accounts: {
+                  evm: {
                     [blockchainName]: [e.data.addr],
                   },
                 },
                 e.data.exp
               );
+              if (e.data?.isAccountChanged) {
+                postMessage({
+                  originChain: CHAIN.ETHEREUM,
+                  type: 'BLOCTO_SDK:ACCOUNT_CHANGED',
+                });
+              }
+              addEventListener('message', this.#onAccountChanged);
               resolve([e.data.addr]);
             }
 
@@ -606,7 +624,7 @@ export default class EthereumProvider
     this.#checkNetworkMatched();
     const { blockchainName, sessionKey } = await this.#getBloctoProperties();
     const { accounts } = await this.bloctoApi<{ accounts: [] }>(`/accounts`);
-    setChainAddress(sessionKey, blockchainName, accounts);
+    setEvmAddress(sessionKey, blockchainName, accounts);
     return accounts;
   }
 
@@ -675,14 +693,9 @@ export default class EthereumProvider
     if (!targetChainId) {
       throw ethErrors.rpc.invalidParams();
     }
-    const {
-      walletServer,
-      blockchainName,
-      sessionKey,
-      switchableNetwork,
-      supportNetworkList,
-    } = await this.#getBloctoProperties();
-    const oldAccount = getChainAddress(sessionKey, blockchainName)?.[0];
+    const { walletServer, blockchainName, sessionKey, switchableNetwork } =
+      await this.#getBloctoProperties();
+    const oldAccount = getEvmAddress(sessionKey, blockchainName)?.[0];
     const oldChainId = parseChainId(this.chainId);
     const newChainId = parseChainId(targetChainId);
     if (oldChainId === newChainId) {
@@ -748,8 +761,7 @@ export default class EthereumProvider
                   sessionKey,
                   {
                     code: e.data?.code,
-                    connected: true,
-                    accounts: {
+                    evm: {
                       [switchableNetwork[newChainId].name]: [e.data.addr],
                     },
                   },
@@ -775,16 +787,7 @@ export default class EthereumProvider
                 this.eventListeners?.chainChanged.forEach((listener) =>
                   listener(this.chainId)
                 );
-                Object.values(supportNetworkList).map(
-                  ({ name, blocto_service_environment }) => {
-                    if (
-                      sessionKey ===
-                      ETH_SESSION_KEY_MAPPING[blocto_service_environment]
-                    ) {
-                      removeChainAddress(sessionKey, name);
-                    }
-                  }
-                );
+                removeAllEvmAddress(sessionKey);
                 this.eventListeners?.disconnect.forEach((listener) =>
                   listener(ethErrors.provider.disconnected())
                 );
@@ -871,17 +874,9 @@ export default class EthereumProvider
     if (existedSDK && existedSDK.isBlocto) {
       return existedSDK.disconnect();
     }
-    const { sessionKey, supportNetworkList } =
-      await this.#getBloctoProperties();
-    Object.values(supportNetworkList).map(
-      ({ name, blocto_service_environment }) => {
-        if (
-          sessionKey === ETH_SESSION_KEY_MAPPING[blocto_service_environment]
-        ) {
-          removeChainAddress(sessionKey, name);
-        }
-      }
-    );
+    const { sessionKey } = await this.#getBloctoProperties();
+    removeAllEvmAddress(sessionKey);
+    removeEventListener('message', this.#onAccountChanged);
     this.eventListeners?.disconnect.forEach((listener) =>
       listener(ethErrors.provider.disconnected())
     );
@@ -897,11 +892,14 @@ export default class EthereumProvider
     if (networkList?.length) {
       const listToAdd = networkList.map(({ chainId, rpcUrls }) => {
         if (!chainId) throw ethErrors.rpc.invalidParams('Empty chainId');
-        if (!rpcUrls?.length)
-          throw ethErrors.rpc.invalidParams('Empty rpcUrls');
+        const parsedChainId: `${number}` = `${parseChainId(chainId)}`;
+        // skip if chainId already exists
+        if (this._blocto.switchableNetwork[parsedChainId]) return null;
+        const parsedRpc = rpcUrls?.[0] || ETH_RPC_LIST[parsedChainId];
+        if (!parsedRpc) throw ethErrors.rpc.invalidParams('rpcUrls required');
         return this.#addToSwitchable({
-          chainId: `${parseChainId(chainId)}`,
-          rpcUrls,
+          chainId: parsedChainId,
+          rpcUrls: [parsedRpc],
         });
       });
       return Promise.all(listToAdd).then(() => null);